reviews separated into own component and useRef added for creating review
