+++ frontend/src/components/Navigation/Navigation.css
@@ -4,8 +4,9 @@
     position: fixed;
     top: 0;
     display: flex;
-    width: 88%;
+    width: 100%;
     height: 78.3px;
+    padding: 0px 50px;
     background-color: white;
     justify-content: space-between;
 }
@@ -17,7 +18,7 @@
     align-items: center;
     /* width: 33%; */
     position: relative;
-    left: 1%;
+    /* left: 1%; */
     gap: 4px;
     /* top: 2.25%; */
 }
